requirements.txt
numpy == 1.21.4
pandas == 1.3.4
pickleshare == 0.7.5
python == 3.7.11
pytorch == 1.7.1
pytz == 2021.3
pyzmq == 19.0.2
requests = 2.26.0
scikit-learn == 1.0.1
scipy == 1.7.3
sklearn ==0.0
torch-cluster == 1.5.9
torch-geometric == 2.0.2
torch-scatter == 2.0.7
torch-sparse == 0.6.9
torchaudio == 0.7.2
torchvision == 0.8.2
tqdm == 4.62.3
urllib3 == 1.26.7
wheel == 0.37.0
zipp == 3.6.0

